+++ blocks/header/header.css
@@ -40,6 +40,7 @@ header nav .nav-brand p {
 
 header nav .nav-brand img {
   width: 128px;
+  height: auto;
 }
 
 /* hamburger */
