+++ blocks/video/video.css
@@ -1,0 +1,3 @@
+.video video {
+    max-width: 100%;
+}
